+++ GalagaRacerFX/src/application/Entities/GroupedEntities/Emitters/Emitter.java
@@ -1,7 +1,9 @@
-package application;
+package application.Entities.GroupedEntities.Emitters;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import application.Entities.GroupedEntities.Particle;
 import javafx.geometry.Point3D;
 import javafx.scene.Group;
