+++ GalagaRacerFX/src/application/Main.java
@@ -2,6 +2,14 @@
 package application;
 import java.io.File;
 
+import application.Entities.Lane;
+import application.Entities.PlayerShip;
+import application.Entities.GroupedEntities.StaticEntity;
+import application.Light.LightHandler;
+import application.Light.LightInstance;
+import application.State.ControlShip;
+import application.State.GameCamera;
+import application.State.GameState;
 import application.UI.HUD;
 import application.UI.Menus;
 import application.UI.PauseScreen;
