+++ GalagaRacerFX/src/application/State/ControlShip.java
@@ -1,9 +1,16 @@
-package application;
+package application.State;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
+import application.Main;
+import application.Entities.Lane;
+import application.Entities.PlayerShip;
+import application.Entities.GroupedEntities.Particle;
+import application.Entities.GroupedEntities.StaticEntity;
+import application.Entities.GroupedEntities.Emitters.Emitter;
+import application.Entities.GroupedEntities.Emitters.ThrustEmitter;
 import application.UI.HUD;
 import application.UI.PauseScreen;
 import javafx.animation.AnimationTimer;
@@ -89,7 +96,7 @@ public class ControlShip {
 		this.currSpeed = minSpeed;
 		this.shiftProp = shiftProp;
 		
-		// handle keyboard -- move to a seperate function at some point maybe!
+		// handle keyboard -- move to a separate function at some point maybe!
         scene.setOnKeyPressed(event -> handleKeyPress(event.getCode()));
         scene.setOnKeyReleased(event -> handleKeyRelease(event.getCode()));
