+++ GalagaRacerFX/src/application/State/GameState.java
@@ -1,4 +1,4 @@
-package application;
+package application.State;
 
 public enum GameState {
 	RUNNING,
